Added logic to process minutes prediction from rotowire

+++ bayesian-model/src/main/java/nba/minutedistribution/ZeroMinutesModel.java
@@ -4,7 +4,7 @@ import org.apache.commons.math3.util.FastMath;
 
 public class ZeroMinutesModel {
 
-    public static double zeroMinutesProb(int pmin, int starter, int hasOt){
+    public static double zeroMinutesProb(double pmin, int starter, int hasOt){
         double exp = Math.exp(0.59926 +
                 -0.22910  * pmin +
                 (starter == 1 ? -15.76094 : 0) +
